Move GithubSvgIcon, Image, SVG, UnknownIcon into Base directory
