+++ packages/react-web3-icons/src/components/Base/Image.tsx
