+++ packages/react-web3-icons/src/components/Base/SVG.tsx
