+++ packages/react-web3-icons/src/components/index.ts
@@ -1,3 +1,3 @@
+export * from "./Base/SVG";
 export * from "./StaticIcon";
-export * from "./SVG";
 export * from "./Web3Icon";
